Refactor project and settings.
